make all size_t

+++ src/RcppExports.cpp
@@ -6,31 +6,31 @@
 using namespace Rcpp;
 
 // orthoKMeansTrainCpp
-List orthoKMeansTrainCpp(Rcpp::NumericMatrix x, unsigned int rounds, unsigned int k, unsigned int iter, unsigned int initType, bool verbose);
+List orthoKMeansTrainCpp(Rcpp::NumericMatrix x, size_t rounds, size_t k, size_t iter, size_t initType, bool verbose);
 RcppExport SEXP yakmoR_orthoKMeansTrainCpp(SEXP xSEXP, SEXP roundsSEXP, SEXP kSEXP, SEXP iterSEXP, SEXP initTypeSEXP, SEXP verboseSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
     Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type x(xSEXP);
-    Rcpp::traits::input_parameter< unsigned int >::type rounds(roundsSEXP);
-    Rcpp::traits::input_parameter< unsigned int >::type k(kSEXP);
-    Rcpp::traits::input_parameter< unsigned int >::type iter(iterSEXP);
-    Rcpp::traits::input_parameter< unsigned int >::type initType(initTypeSEXP);
+    Rcpp::traits::input_parameter< size_t >::type rounds(roundsSEXP);
+    Rcpp::traits::input_parameter< size_t >::type k(kSEXP);
+    Rcpp::traits::input_parameter< size_t >::type iter(iterSEXP);
+    Rcpp::traits::input_parameter< size_t >::type initType(initTypeSEXP);
     Rcpp::traits::input_parameter< bool >::type verbose(verboseSEXP);
     __result = Rcpp::wrap(orthoKMeansTrainCpp(x, rounds, k, iter, initType, verbose));
     return __result;
 END_RCPP
 }
 // orthoKMeansPredictCpp
-List orthoKMeansPredictCpp(NumericMatrix x, std::vector <NumericMatrix> centers, unsigned int nf, unsigned int k, bool verbose);
+List orthoKMeansPredictCpp(NumericMatrix x, std::vector <NumericMatrix> centers, size_t nf, size_t k, bool verbose);
 RcppExport SEXP yakmoR_orthoKMeansPredictCpp(SEXP xSEXP, SEXP centersSEXP, SEXP nfSEXP, SEXP kSEXP, SEXP verboseSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
     Rcpp::traits::input_parameter< NumericMatrix >::type x(xSEXP);
     Rcpp::traits::input_parameter< std::vector <NumericMatrix> >::type centers(centersSEXP);
-    Rcpp::traits::input_parameter< unsigned int >::type nf(nfSEXP);
-    Rcpp::traits::input_parameter< unsigned int >::type k(kSEXP);
+    Rcpp::traits::input_parameter< size_t >::type nf(nfSEXP);
+    Rcpp::traits::input_parameter< size_t >::type k(kSEXP);
     Rcpp::traits::input_parameter< bool >::type verbose(verboseSEXP);
     __result = Rcpp::wrap(orthoKMeansPredictCpp(x, centers, nf, k, verbose));
     return __result;
